backend/requirements.txt
annotated-types==0.7.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
diff-match-patch==20241021
Django==5.1.5
django-autocomplete-light==3.11.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==3.2.3
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.8
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
markdown2==2.5.3
mdurl==0.1.2
mutagen==1.47.0
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
tablib==3.7.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
zope.interface==7.2
urllib3==2.3.0
whitenoise==6.8.2
zope.interface==7.2
